requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=-
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   groq
    #   httpx
    #   starlette
    #   watchfiles
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via groq
dnspython==2.6.1
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.112.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
filelock==3.16.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via
    #   huggingface-hub
    #   torch
groq==0.11.0
    # via langchain-groq
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via
    #   fastapi
    #   groq
    #   langsmith
huggingface-hub==0.24.7
    # via
    #   langchain-huggingface
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   fastapi
    #   torch
joblib==1.4.2
    # via scikit-learn
jq==1.8.0
    # via -r requirements.in
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.0
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.0
    # via -r requirements.in
langchain-core==0.3.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-groq
    #   langchain-huggingface
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
langchain-groq==0.2.0
    # via -r requirements.in
langchain-huggingface==0.1.0
    # via -r requirements.in
langchain-text-splitters==0.3.0
    # via langchain
langgraph==0.2.21
    # via -r requirements.in
langgraph-checkpoint==1.0.9
    # via langgraph
langsmith==0.1.120
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.22.0
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   transformers
pillow==10.4.0
    # via sentence-transformers
pydantic==2.9.1
    # via
    #   fastapi
    #   groq
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.23.3
    # via pydantic
pydantic-settings==2.5.2
    # via langchain-community
pygments==2.18.0
    # via rich
pymongo==4.8.0
    # via -r requirements.in
pypdf==4.3.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
    #   uvicorn
regex==2024.9.11
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   transformers
rich==13.8.1
    # via typer
safetensors==0.4.5
    # via transformers
scikit-learn==1.5.2
    # via sentence-transformers
scipy==1.14.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.1.0
    # via langchain-huggingface
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   groq
    #   httpx
sqlalchemy==2.0.34
    # via
    #   langchain
    #   langchain-community
starlette==0.38.5
    # via fastapi
sympy==1.13.2
    # via torch
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via
    #   langchain-huggingface
    #   transformers
torch==2.4.1
    # via sentence-transformers
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.44.2
    # via
    #   langchain-huggingface
    #   sentence-transformers
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   groq
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.30.6
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
yarl==1.11.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
